Moving redacted_in_out_flows into it's own file
