+++ flowmachine/flowmachine/features/location/redacted_flows.py
@@ -5,7 +5,7 @@
 
 from flowmachine.core import Query
 from flowmachine.features import Flows
-from flowmachine.features.location.flows import FlowLike, BaseInOutFlow
+from flowmachine.features.location.flows import FlowLike
 from flowmachine.features.location.redacted_location_metric import (
     RedactedLocationMetric,
 )
@@ -27,15 +27,3 @@ class RedactedFlows(RedactedLocationMetric, FlowLike, Query):
         # self.spatial_unit is used in self._geo_augmented_query
         self.spatial_unit = flows.spatial_unit
         super().__init__()
-
-
-class RedactedInOutFlow(RedactedLocationMetric, BaseInOutFlow, Query):
-    """
-    An object representing the summation
-    """
-
-    def __init__(self, *, flows: BaseInOutFlow):
-        self.redaction_target = flows
-        # self.spatial_unit is used in self._geo_augmented_query
-        self.spatial_unit = flows.spatial_unit
-        super().__init__(flows.flow)
